+++ CHANGES.md
@@ -1,6 +1,34 @@
 
 ## Unreleased
 
+* use `WGS84` as default CRS for `TileMatrixSet.feature` GeoJSON response (as per specification)
+* add `geographic_crs` option for `TileMatrixSet.feature` method
+* update non-WGS84 CRS notation in `TileMatrixSet.feature` GeoJSON response
+
+    ```python
+    # before
+    "properties": {
+        "crs": {
+            "type": "EPSG",
+            "properties": {"code": 3857},
+        }
+    }
+
+    # now
+    "properties": {
+        "crs": {
+            "type": "name",
+            "properties": {"name": "http://www.opengis.net/def/crs/EPSG/0/3857"},
+        }
+        # or
+        "crs": {
+            "type": "wkt",
+             "properties": {"wkt": "..."}},
+        }
+    }
+    ```
+
+* rename `grid_name -> tms` and `grid_crs -> tms_crs` property names in `TileMatrixSet.feature` GeoJSON response
 * remove python 3.8 support
 * check tile's zoom against TMS's `maxzoom` in `TileMatrixSet.is_valid` and add `strict=True|False` options